Simplify downloadFontFaceAsset and move to utils.
